+++ release/ewoh-0.6.0-rc2/ewoh-spark-app/server/modules/oee/oee.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { OeeController } from './oee.controller';
+import { OeeService } from './oee.service';
+
+@Module({
+  controllers: [OeeController],
+  providers: [OeeService],
+  exports: [OeeService],
+})
+export class OeeModule {}